irc: remove undirected echo, add directed ping
Stop repeating everything anyone says because that's not going to fly in
an actual channel. Instead, say a fixed phrase ("beep boop") whenever
the bot is addressed directly, either via direct /msg or via a
nick-prefixed message in a channel ("botname: hello"). This can be used
as a ping to check if the bot is working or not.